Create data-dual-full.html template for isvhsopen
This template is for displaying two datapoints at once. Visually, datapoint1 is the primary datapoint, and datapoint2 is the secondary.
The primary use for this is for the new isvhsopen.com. datapoint1.value will be either 'closed' or 'open' and datapoint2.value might be 'until 12:45'. CSS formatting is specifically designed for this to look good, but other datapoint values will still work fine.

Note that the "as of" tooltip only shows the last updated time of datapoint1.